Move EC key operations out of general `utils.ts` and into `keys.ts`
