+++ src/keys.ts
@@ -1,8 +1,9 @@
 
-import { ECPair, address as baddress, networks } from 'bitcoinjs-lib'
+import { ECPair, address, networks } from 'bitcoinjs-lib'
 import { randomBytes } from './encryption/cryptoRandom'
 import { createSha2Hash } from './encryption/sha2Hash'
 import { createHashRipemd160 } from './encryption/hashRipemd160'
+import { config } from './config'
 
 /**
  * 
@@ -34,8 +35,8 @@ export async function publicKeyToAddress(publicKey: string | Buffer) {
   const publicKeyHash160 = await createHashRipemd160().digest(
     await sha2Hash.digest(publicKeyBuffer)
   )
-  const address = baddress.toBase58Check(publicKeyHash160, networks.bitcoin.pubKeyHash)
-  return address
+  const result = address.toBase58Check(publicKeyHash160, networks.bitcoin.pubKeyHash)
+  return result
 }
 
 /**
@@ -45,4 +46,54 @@ export function getPublicKeyFromPrivate(privateKey: string | Buffer) {
   const privateKeyBuffer = Buffer.isBuffer(privateKey) ? privateKey : Buffer.from(privateKey, 'hex')
   const keyPair = ECPair.fromPrivateKey(privateKeyBuffer)
   return keyPair.publicKey.toString('hex')
+}
+
+/**
+ * Time
+ * @private
+ * @ignore
+ */
+export function hexStringToECPair(skHex: string): ECPair.ECPairInterface {
+  const ecPairOptions = {
+    network: config.network.layer1,
+    compressed: true
+  }
+
+  if (skHex.length === 66) {
+    if (skHex.slice(64) !== '01') {
+      throw new Error('Improperly formatted private-key hex string. 66-length hex usually '
+                      + 'indicates compressed key, but last byte must be == 1')
+    }
+    return ECPair.fromPrivateKey(Buffer.from(skHex.slice(0, 64), 'hex'), ecPairOptions)
+  } else if (skHex.length === 64) {
+    ecPairOptions.compressed = false
+    return ECPair.fromPrivateKey(Buffer.from(skHex, 'hex'), ecPairOptions)
+  } else {
+    throw new Error('Improperly formatted private-key hex string: length should be 64 or 66.')
+  }
+}
+
+/**
+ * 
+ * @ignore
+ */
+export function ecPairToHexString(secretKey: ECPair.ECPairInterface) {
+  const ecPointHex = secretKey.privateKey.toString('hex')
+  if (secretKey.compressed) {
+    return `${ecPointHex}01`
+  } else {
+    return ecPointHex
+  }
+}
+
+/**
+ * Creates a bitcoin address string from an ECPair
+ * @private
+ * @ignore
+ */
+export async function ecPairToAddress(keyPair: ECPair.ECPairInterface) {
+  const sha2Hash = await createSha2Hash()
+  const sha256 = await sha2Hash.digest(keyPair.publicKey)
+  const hash160 = await createHashRipemd160().digest(sha256)
+  return address.toBase58Check(hash160, keyPair.network.pubKeyHash)
 }
