+++ src/profiles/profileTokens.ts
@@ -1,8 +1,8 @@
 import { ECPair } from 'bitcoinjs-lib'
 import { decodeToken, SECP256K1Client, TokenSigner, TokenVerifier } from 'jsontokens'
 import { TokenInterface } from 'jsontokens/lib/decode'
-import { nextYear, makeUUID4, ecPairToAddress } from '../utils'
-
+import { nextYear, makeUUID4 } from '../utils'
+import { ecPairToAddress } from '../keys'
 
 /**
   * Signs a profile token
