+++ src/utils.ts
@@ -1,5 +1,3 @@
-import { ECPair, address } from 'bitcoinjs-lib'
-import { config } from './config'
 import { Logger } from './logger'
 import { 
   BadPathError, 
@@ -10,8 +8,6 @@ import {
   PayloadTooLargeError, 
   ValidationError
 } from './errors'
-import { createSha2Hash } from './encryption/sha2Hash'
-import { createHashRipemd160 } from './encryption/hashRipemd160'
 
 
 /**
@@ -104,56 +100,6 @@ export function isLaterVersion(v1: string, v2: string) {
     }
   }
   return true
-}
-
-/**
- * Time
- * @private
- * @ignore
- */
-export function hexStringToECPair(skHex: string): ECPair.ECPairInterface {
-  const ecPairOptions = {
-    network: config.network.layer1,
-    compressed: true
-  }
-
-  if (skHex.length === 66) {
-    if (skHex.slice(64) !== '01') {
-      throw new Error('Improperly formatted private-key hex string. 66-length hex usually '
-                      + 'indicates compressed key, but last byte must be == 1')
-    }
-    return ECPair.fromPrivateKey(Buffer.from(skHex.slice(0, 64), 'hex'), ecPairOptions)
-  } else if (skHex.length === 64) {
-    ecPairOptions.compressed = false
-    return ECPair.fromPrivateKey(Buffer.from(skHex, 'hex'), ecPairOptions)
-  } else {
-    throw new Error('Improperly formatted private-key hex string: length should be 64 or 66.')
-  }
-}
-
-/**
- * 
- * @ignore
- */
-export function ecPairToHexString(secretKey: ECPair.ECPairInterface) {
-  const ecPointHex = secretKey.privateKey.toString('hex')
-  if (secretKey.compressed) {
-    return `${ecPointHex}01`
-  } else {
-    return ecPointHex
-  }
-}
-
-/**
- * Time
- * @private
- * @ignore
- */
-export async function ecPairToAddress(keyPair: ECPair.ECPairInterface) {
-  const sha2Hash = await createSha2Hash()
-  const sha256 = await sha2Hash.digest(keyPair.publicKey)
-  const hash160 = await createHashRipemd160().digest(sha256)
-  return address.toBase58Check(hash160, keyPair.network.pubKeyHash)
 }
 
 /**
